[SkyMarshal] Explicitly static_cast hashes to int64
Topic: lcm-hash-cast
Reviewers: hayk,bradley,abe,peter
Relative: eigenlcm-fix
GitOrigin-RevId: 47e296c7c7eff2540a57cef2d2cf5ca9bbd02ab6

+++ test/symforce_function_codegen_test_data/symengine/codegen_python_test_data/cpp/lcmtypes/codegen_python_test/constants_t.hpp
@@ -144,7 +144,7 @@ int constants_t::getEncodedSize() const
 
 constexpr int64_t constants_t::getHash()
 {
-    return _computeHash(NULL);
+    return static_cast<int64_t>(_computeHash(NULL));
 }
 
 constexpr constants_t::type_name_array_t* constants_t::getTypeNameArrayPtr() {
